Fix CC processing of dates and new format.

+++ sec_certs_page/cc/tasks.py
@@ -6,6 +6,7 @@ from operator import itemgetter
 
 import sentry_sdk
 from celery.utils.log import get_task_logger
+from flask import current_app
 from pkg_resources import get_distribution
 from sec_certs.dataset.common_criteria import CCDataset
 
@@ -28,7 +29,13 @@ def update_data():  # pragma: no cover
     start = datetime.now()
     paths = make_dataset_paths("cc")
 
-    dset = CCDataset({}, paths["dset_path"], "dataset", "Description")
+    if current_app.config["CC_SKIP_UPDATE"] and paths["output_path"].exists():
+        dset = CCDataset.from_json(paths["output_path"])
+        dset.root_dir = paths["dset_path"]
+        dset.set_local_paths()
+    else:
+        dset = CCDataset({}, paths["dset_path"], "dataset", "Description")
+
     if not dset.auxillary_datasets_dir.exists():
         dset.auxillary_datasets_dir.mkdir(parents=True)
     if paths["cve_path"].exists():
@@ -38,18 +45,20 @@ def update_data():  # pragma: no cover
 
     try:
         with sentry_sdk.start_span(op="cc.all", description="Get full CC dataset"):
-            with sentry_sdk.start_span(op="cc.get_certs", description="Get certs from web"):
-                dset.get_certs_from_web(update_json=False)
-            with sentry_sdk.start_span(op="cc.download_pdfs", description="Download pdfs"):
-                dset.download_all_pdfs(update_json=False)
-            with sentry_sdk.start_span(op="cc.convert_pdfs", description="Convert pdfs"):
-                dset.convert_all_pdfs(update_json=False)
-            with sentry_sdk.start_span(op="cc.analyze", description="Analyze certificates"):
-                dset.analyze_certificates(update_json=False)
-            with sentry_sdk.start_span(op="cc.maintenance_updates", description="Process maintenance updates"):
-                dset.process_maintenance_updates()
-            with sentry_sdk.start_span(op="cc.write_json", description="Write JSON"):
-                dset.to_json(paths["output_path"])
+            if not current_app.config["CC_SKIP_UPDATE"] or not paths["output_path"].exists():
+                with sentry_sdk.start_span(op="cc.get_certs", description="Get certs from web"):
+                    dset.get_certs_from_web(update_json=False)
+                with sentry_sdk.start_span(op="cc.download_pdfs", description="Download pdfs"):
+                    dset.download_all_pdfs(update_json=False)
+                with sentry_sdk.start_span(op="cc.convert_pdfs", description="Convert pdfs"):
+                    dset.convert_all_pdfs(update_json=False)
+                with sentry_sdk.start_span(op="cc.analyze", description="Analyze certificates"):
+                    dset.analyze_certificates(update_json=False)
+                with sentry_sdk.start_span(op="cc.maintenance_updates", description="Process maintenance updates"):
+                    dset.process_maintenance_updates()
+                with sentry_sdk.start_span(op="cc.write_json", description="Write JSON"):
+                    dset.to_json(paths["output_path"])
+
             with sentry_sdk.start_span(op="cc.move", description="Move files"):
                 for cert in dset:
                     if cert.state.report_pdf_path and cert.state.report_pdf_path.exists():
